Resolved some issues + created API Documentation
+ Resolved issue that allowed users to submit a friend requests to an existing friend
+ Resolved issue that allow users to delete friendships requests that didn't exist
+ resolved issue not showing search results
+ Created new function that checks for existing friendships
+ Made variable name changes
+ Created API Documentation for Login, Register and Friendships

+++ API/controllers/friendshipController.js
@@ -4,39 +4,30 @@ const sqlConfig = require('../config');
 const { stringify } = require('querystring');
 
 //sends a friend request
-function sendRequest(requester, requestee, res){
-    //take in the requester and Requestee's Account ID
-    const friendship = {
-        Requester: requester,
-        requestee: requestee,
-        Status: "pending"
-    }
-
-    //submit the friendship to the database
+function sendRequest(requesterID, requesteeID, res){
     try{
         sql.connect(sqlConfig.returnServerConfig()).then(async function(){
-            //check DB for an existing friendrequest
-            var result = await sql.query`Select * FROM Friendships
-                                           Where RequesterID = ${friendship.Requester}
-                                           AND AddresseeID = ${friendship.requestee}
-                                           
-                                           OR RequesterID = ${friendship.requestee} 
-                                           AND AddresseeID = ${friendship.Requester}`
-            
-            //if there isn't an existing friend request, add friendship into the database as pending
-            if(result.recordsets == 0){
+            const existingFriendship = await checkForExistingFriendships(requesterID, requesteeID);
+
+            //check DB for an existing friendship or friend request
+            if(!existingFriendship){
+                //if there isn't an existing friend request, add friendship into the database as pending
                 result = await sql.query`INSERT INTO Friendships (RequesterID, AddresseeID, Status)
-                                         Values (${friendship.Requester}, ${friendship.requestee}, 'Pending')`
+                Values (${requesterID}, ${requesteeID}, 'Pending')`
+
                 return res.status(200).json({
                     Message: "OK"
                 });
-            } else {
-                return res.status(400).json({
-                    Message: "Friendship already exists"
-                })
-            }
+            } 
+
+            //if friendship / friend request already exists return 401
+            return res.status(401).json({
+                Message: "Friendship already exists"
+            });
         })
-    } catch{
+    //catch any errors
+    } catch(err){
+        console.dir(err);
         return res.status(400).json({
             Message: "Server Error"
         })
@@ -44,20 +35,31 @@ function sendRequest(requester, requestee, res){
 }
 
 //accept a friend request
-function acceptRequest(currentUserID, OtherUserID, res){
+async function acceptRequest(currentUserID, requesterID, res){
     try{
-        
+        const existingFriendship = await checkForExistingFriendships(currentUserID, requesterID);
+        //connect to DB
         sql.connect(sqlConfig.returnServerConfig()).then(async function(){
-            const result = await sql.query`UPDATE Friendships
-                                           SET Status = 'Accepted'
-                                           WHERE RequesterID = ${OtherUserID}
-                                           AND AddresseeID = ${currentUserID}`
-            return res.status(200).json({
-                Message: "OK"
-            });
-        });
+            // if friendship exists, update database and return OK code
+            if(existingFriendship){
+                const result = await sql.query`UPDATE Friendships
+                SET Status = 'Active'
+                WHERE RequesterID = ${requesterID}
+                AND AddresseeID = ${currentUserID}`
+                
+                return res.status(200).json({
+                    Message: "OK"
+                });
+            }
 
-    }catch{
+            //if friendship doesn't exists return error 
+            return res.status(401).json({
+                Message: "Friendship not found!"
+            })
+        });
+    //catch any errors
+    }catch(err){
+        console.dir(err);
         return res.status(400).json({
             Message: "Server Error"
         })
@@ -65,21 +67,34 @@ function acceptRequest(currentUserID, OtherUserID, res){
 }
 
 //delete a friendship or friendship request
-function deleteFriendship(currentUserID, otherUserID, res){
+async function deleteFriendship(currentUserID, otherUserID, res){
     try{
-        //Delete a friendship or friendship request that has been sent by another user.
-        sql.connect(sqlConfig.returnServerConfig()).then(async function(){
-            const result = await sql.query`DELETE FROM Friendships
-                                           WHERE RequesterID = ${otherUserID}
-                                           AND AddresseeID = ${currentUserID}
-                                           OR
-                                           AddresseeID = ${otherUserID}
-                                           AND RequesterID = ${currentUserID}`
-            return res.status(200).json({
-                Message: "OK"
-            });
-        });
-    }catch{
+        //check if friendship exists in the database
+        const existingFriendship = await checkForExistingFriendships(currentUserID, otherUserID);
+ 
+        //check if friendship exists
+        if(existingFriendship){
+            //Delete a friendship or friendship request that has been sent by another user.
+            sql.connect(sqlConfig.returnServerConfig()).then(async function(){
+                const result = await sql.query`DELETE FROM Friendships
+                                            WHERE RequesterID = ${otherUserID}
+                                            AND AddresseeID = ${currentUserID}
+                                            OR
+                                            AddresseeID = ${otherUserID}
+                                            AND RequesterID = ${currentUserID}`
+                return res.status(200).json({
+                    Message: "OK"
+                })
+            })
+        } else {
+            //return 401 if friendship isn't found
+            return res.status(401).json({
+                Message: "Friendship not found"
+            })
+        }
+    //catch and print any errors encountered 
+    }catch(err){
+        console.dir(err);
         return res.status(400).json({
             Message: "Server Error"
         });
@@ -90,13 +105,24 @@ function deleteFriendship(currentUserID, otherUserID, res){
 function search(DisplayName, res){
     try{
         sql.connect(sqlConfig.returnServerConfig()).then(async function(){
-            const result = await sql.query('SELECT * FROM Accounts WHERE DisplayName LIKE \'%' + DisplayName + '%\' ')
-
-            return res.status(200).json({
-                Message: "OK",
-                result
+            //select results similar to the input display name entered
+            const result = await sql.query('SELECT AccountID, Email, DisplayName FROM Accounts WHERE DisplayName LIKE \'%' + DisplayName + '%\' ')
+            
+            const userList = result.recordsets;
+            //return any results found
+            if(result.rowsAffected > 0){
+                return res.status(200).json({
+                    Message: "OK",
+                    userList
+                });
+            }
+            //return 401, if no results found
+            return res.status(401).json({
+                Message: "No results found!"
             });
+
         });
+    //catch any errors
     } catch{
         return res.status(400).json({
             Message: "Server Error"
@@ -105,7 +131,7 @@ function search(DisplayName, res){
 }
 
 //return a list of users friends or friendrequests
-function returnFriendsList(currentUserID, res){
+function returnFriendsList(currentUserID, status, res){
     try{
         //select all users from the friendships table that match the users AccountID
         //client side provides the requested status such as pending or accepted
@@ -113,18 +139,62 @@ function returnFriendsList(currentUserID, res){
             const result = await sql.query`SELECT * FROM friendships
 
                                            WHERE RequesterID = ${currentUserID}
-                                           OR AddresseeID = ${currentUserID}`
+                                           AND Status = ${status}
+                                           OR AddresseeID = ${currentUserID}
+                                           AND Status = ${status}`
 
-            return res.status(200).json({
-                Message: "OK",
-                result
+            const friendships = result.recordsets;
+            //return the list of users
+            if(result.rowsAffected > 0){
+                return res.status(200).json({
+                    Message: "OK",
+                    friendships
+                });
+            }
+            
+            //if not friendships are found of specified Type, return 401.
+            return res.status(401).json({
+                Message: "No results found!"
             });
         });
-    }catch{
+        
+    //catch and print any errors
+    }catch(err){
+        console.dir(err);
         return res.status(400).json({
             Message: "Server Error"
         });
     }
+}
+
+//checks for an existing friendship in the database, returns true if results are found
+async function checkForExistingFriendships(currentUserID, otherUserID){
+    //submit the friendship to the database
+    try{
+        await sql.connect(sqlConfig.returnServerConfig());
+            //check DB for an existing friendrequest
+        const query = `Select TOP 1 * FROM Friendships
+                                           Where RequesterID = ${currentUserID}
+                                           AND AddresseeID = ${otherUserID}
+                                           
+                                           OR RequesterID = ${otherUserID} 
+                                           AND AddresseeID = ${currentUserID}`
+
+        const result = await sql.query(query);
+
+        //if there is an existing friend request return true
+        if(result.recordset.length > 0){
+            return Promise.resolve(true);
+        }
+        // //return false if no friendship found
+        return Promise.resolve(false);
+
+    //return false in an error occurs
+    } catch(err){
+        console.dir(err);
+        return Promise.resolve(false);
+    }
+
 }
 
 module.exports = {
